fix: android files incorrect app name, bumped some packages
