+++ example/android/app/src/main/jni/MainApplicationTurboModuleManagerDelegate.h
@@ -11,7 +11,7 @@ class MainApplicationTurboModuleManagerDelegate
  public:
   // Adapt it to the package you used for your Java class.
   static constexpr auto kJavaDescriptor =
-      "Lcom/rndiffapp/newarchitecture/modules/MainApplicationTurboModuleManagerDelegate;";
+      "Lcom/example/newarchitecture/modules/MainApplicationTurboModuleManagerDelegate;";
   static jni::local_ref<jhybriddata> initHybrid(jni::alias_ref<jhybridobject>);
   static void registerNatives();
   std::shared_ptr<TurboModule> getTurboModule(
