+++ example/package.json
@@ -18,10 +18,10 @@
     "@reduxjs/toolkit": "^1.7.1",
     "react": "17.0.2",
     "react-native": "0.68.4",
-    "react-native-gesture-handler": "2.1.1",
-    "react-native-reanimated": "^2.10.0",
+    "react-native-gesture-handler": "2.8.0",
+    "react-native-reanimated": "^3.0.0-rc.9",
     "react-native-safe-area-context": "^3.3.2",
-    "react-native-screens": "^3.14.0",
+    "react-native-screens": "^3.18.2",
     "react-redux": "^7.2.6"
   },
   "devDependencies": {
